fix(pr): delete temporary layers before discovery
