+++ internal/controllers/terraformpullrequest/layer.go
@@ -5,6 +5,9 @@ import (
 	"fmt"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/selection"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	configv1alpha1 "github.com/padok-team/burrito/api/v1alpha1"
 	"github.com/padok-team/burrito/internal/annotations"
@@ -94,4 +97,24 @@ func generateTempLayers(pr *configv1alpha1.TerraformPullRequest, layers []config
 		list = append(list, new)
 	}
 	return list
+}
+
+func GetLinkedLayers(cl client.Client, pr *configv1alpha1.TerraformPullRequest) ([]configv1alpha1.TerraformLayer, error) {
+	layers := configv1alpha1.TerraformLayerList{}
+	requirement, err := labels.NewRequirement("burrito/managed-by", selection.Equals, []string{pr.Name})
+	if err != nil {
+		return nil, err
+	}
+	selector := labels.NewSelector().Add(*requirement)
+	err = cl.List(context.TODO(), &layers, client.MatchingLabelsSelector{Selector: selector})
+	if err != nil {
+		return nil, err
+	}
+	return layers.Items, nil
+}
+
+func (r *Reconciler) deleteTempLayers(ctx context.Context, pr *configv1alpha1.TerraformPullRequest) error {
+	return r.Client.DeleteAllOf(
+		ctx, &configv1alpha1.TerraformLayer{}, client.InNamespace(pr.Namespace), client.MatchingLabels{"burrito/managed-by": pr.Name},
+	)
 }
